+++ doc/CHANGELOG.md
@@ -29,6 +29,7 @@ See doc/COPYRIGHT.rdoc for more details.
 
 # Changelog
 
+* `#3028` Migration of legacy planning elements doesn't update journals.
 * `#3030` Users preferences for order of comments is ignored on wp comments
 
 ## 3.0.0pre31